ai-news-scraper/requirements.txt
# AI News Scraper 依赖包

feedparser>=6.0.10
requests>=2.31.0
beautifulsoup4>=4.12.0
PyYAML>=6.0
schedule>=1.2.0
